PR for #132 (#138)
* organise memos in one hook

* split useAccounting into useAccountingV2 & useAccountingV3

* Remove useAccounting hook

---------
