+++ libs/dm-hooks/src/index.ts
@@ -1,4 +1,5 @@
-export { default as useAccounting } from './useAccounting';
+export { default as useAccountingV2 } from './useAccountingV2';
+export { default as useAccountingV3 } from './useAccountingV3';
 export { default as useApplicationDetail } from './useApplicationDetail';
 export {
   default as useApplicationList,
@@ -15,6 +16,7 @@ export { useContacts } from './useContacts';
 export { default as useContactUpdate } from './useContactUpdate';
 export { default as useDashboardList } from './useDashboardList';
 export { default as useDefaultTitle } from './useDefaultTitle';
+export { default as useFormattedData } from './useFormattedData';
 export * from './useLinks';
 export { default as useLinksUpdate } from './useLinksUpdate';
 export { default as useMemberCreate } from './useMemberCreate';
